pyproject.toml
[project]
name = "self-improving-agent"
version = "0.1.0"
description = "Self-improving AI agent with Anthropic SDK"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "zhipuai>=2.0.0",
    "groq>=0.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
http = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
agent = "agent.main:cli_main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
